Add test for k8sagent package;
