+++ caas/scripts.go
@@ -15,8 +15,6 @@ cp /opt/jujud $JUJU_TOOLS_DIR/jujud
 `[1:]
 
 	// K8sAgentStartUpSh is the start script for in-pod style k8s agents.
-	// TODO(ycliuhw): add a feature flag to enable this new style k8s agents.
-	// - Ensure to use Jujuc for hook tool symlinks but non fallback to jujud because no JujuD at all.
 	K8sAgentStartUpSh = `
 export JUJU_DATA_DIR=%[1]s
 export JUJU_TOOLS_DIR=$JUJU_DATA_DIR/%[2]s
