--- cmd/jujud/hooktool/hooktool.go
@@ -1,142 +1,0 @@
-// Copyright 2020 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package hooktool
-
-import (
-	"crypto/tls"
-	"crypto/x509"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-
-	"github.com/juju/cmd"
-	"github.com/juju/errors"
-	"github.com/juju/loggo"
-	"github.com/juju/names/v4"
-	"github.com/juju/utils/exec"
-
-	"github.com/juju/juju/juju/sockets"
-	// k8sexec "github.com/juju/juju/caas/kubernetes/provider/exec"
-	// // Import the providers.
-	// _ "github.com/juju/juju/provider/all"
-	// "github.com/juju/juju/upgrades"
-	// "github.com/juju/juju/utils/proxy"
-	// "github.com/juju/juju/worker/logsender"
-	"github.com/juju/juju/worker/uniter/runner/jujuc"
-)
-
-var logger = loggo.GetLogger("juju.cmd.jujud.hooktool")
-
-func getenv(name string) (string, error) {
-	value := os.Getenv(name)
-	if value == "" {
-		return "", errors.Errorf("%s not set", name)
-	}
-	return value, nil
-}
-
-func getwd() (string, error) {
-	dir, err := os.Getwd()
-	if err != nil {
-		return "", err
-	}
-	abs, err := filepath.Abs(dir)
-	if err != nil {
-		return "", err
-	}
-	return abs, nil
-}
-
-func getSocket() (sockets.Socket, error) {
-	var err error
-	socket := sockets.Socket{}
-	socket.Address, err = getenv("JUJU_AGENT_SOCKET_ADDRESS")
-	if err != nil {
-		return sockets.Socket{}, err
-	}
-	socket.Network, err = getenv("JUJU_AGENT_SOCKET_NETWORK")
-	if err != nil {
-		return sockets.Socket{}, err
-	}
-
-	// If we are not connecting over tcp, no need for TLS.
-	if socket.Network != "tcp" {
-		return socket, nil
-	}
-
-	caCertFile, err := getenv("JUJU_AGENT_CA_CERT")
-	if err != nil {
-		return sockets.Socket{}, err
-	}
-	caCert, err := ioutil.ReadFile(caCertFile)
-	if err != nil {
-		return sockets.Socket{}, errors.Annotatef(err, "reading %s", caCertFile)
-	}
-	rootCAs := x509.NewCertPool()
-	if ok := rootCAs.AppendCertsFromPEM(caCert); ok == false {
-		return sockets.Socket{}, errors.Errorf("invalid ca certificate")
-	}
-
-	unitName, err := getenv("JUJU_UNIT_NAME")
-	if err != nil {
-		return sockets.Socket{}, err
-	}
-	application, err := names.UnitApplication(unitName)
-	if err != nil {
-		return sockets.Socket{}, errors.Trace(err)
-	}
-	socket.TLSConfig = &tls.Config{
-		RootCAs:    rootCAs,
-		ServerName: application,
-	}
-	return socket, nil
-}
-
-// Main uses JUJU_CONTEXT_ID and JUJU_AGENT_SOCKET_ADDRESS to ask a running unit agent
-// to execute a Command on our behalf. Individual commands should be exposed
-// by symlinking the command name to this executable.
-func Main(commandName string, ctx *cmd.Context, args []string) (code int, err error) {
-	code = 1
-	contextID, err := getenv("JUJU_CONTEXT_ID")
-	if err != nil {
-		return
-	}
-	dir, err := getwd()
-	if err != nil {
-		return
-	}
-	req := jujuc.Request{
-		ContextId:   contextID,
-		Dir:         dir,
-		CommandName: commandName,
-		Args:        args[1:],
-		Token:       os.Getenv("JUJU_AGENT_TOKEN"),
-	}
-	socket, err := getSocket()
-	if err != nil {
-		return
-	}
-	client, err := sockets.Dial(socket)
-	if err != nil {
-		return code, err
-	}
-	defer client.Close()
-	var resp exec.ExecResponse
-	err = client.Call("Jujuc.Main", req, &resp)
-	if err != nil && err.Error() == jujuc.ErrNoStdin.Error() {
-		req.Stdin, err = ioutil.ReadAll(os.Stdin)
-		if err != nil {
-			err = errors.Annotate(err, "cannot read stdin")
-			return
-		}
-		req.StdinSet = true
-		err = client.Call("Jujuc.Main", req, &resp)
-	}
-	if err != nil {
-		return
-	}
-	os.Stdout.Write(resp.Stdout)
-	os.Stderr.Write(resp.Stderr)
-	return resp.Code, nil
-}
