+++ cmd/k8sagent/package_test.go
@@ -1,0 +1,18 @@
+// Copyright 2020 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package main
+
+import (
+	"runtime"
+	"testing"
+
+	gc "gopkg.in/check.v1"
+)
+
+func TestAll(t *testing.T) {
+	if runtime.GOOS != "linux" {
+		t.Skip("k8sagent only runs on Linux")
+	}
+	gc.TestingT(t)
+}
